+++ app/src/main/java/org/simple/clinic/home/overdue/search/OverdueSearchEvent.kt
@@ -5,7 +5,6 @@ import androidx.paging.PagingData
 import org.simple.clinic.activity.permissions.RequiresPermission
 import org.simple.clinic.analytics.NetworkConnectivityStatus
 import org.simple.clinic.home.overdue.OverdueAppointment
-import org.simple.clinic.home.overdue.search.OverdueSearchQueryValidator.Result
 import org.simple.clinic.platform.util.RuntimePermissionResult
 import org.simple.clinic.util.RequiresNetwork
 import org.simple.clinic.widgets.UiEvent
@@ -14,23 +13,13 @@ import java.util.UUID
 
 sealed class OverdueSearchEvent : UiEvent
 
-data class OverdueSearchHistoryLoaded(val searchHistory: Set<String>) : OverdueSearchEvent()
-
-data class OverdueSearchQueryChanged(val searchQuery: String) : OverdueSearchEvent()
-
-data class OverdueSearchQueryValidated(val result: Result) : OverdueSearchEvent()
-
 data class OverdueSearchResultsLoaded(val overdueAppointments: PagingData<OverdueAppointment>) : OverdueSearchEvent()
 
 data class OverduePatientClicked(val patientUuid: UUID) : OverdueSearchEvent()
 
 data class CallPatientClicked(val patientUuid: UUID) : OverdueSearchEvent()
 
-data class OverdueSearchHistoryClicked(val searchQuery: String) : OverdueSearchEvent()
-
 data class OverdueSearchLoadStateChanged(val overdueSearchProgressState: OverdueSearchProgressState) : OverdueSearchEvent()
-
-object OverdueSearchScreenShown : OverdueSearchEvent()
 
 data class OverdueAppointmentCheckBoxClicked(val appointmentId: UUID) : OverdueSearchEvent()