+++ app/src/main/java/org/simple/clinic/home/overdue/search/OverdueSearchInit.kt
@@ -4,20 +4,9 @@ import com.spotify.mobius.First
 import com.spotify.mobius.First.first
 import com.spotify.mobius.Init
 
-class OverdueSearchInit(
-    private val isOverdueSearchV2Enabled: Boolean
-) : Init<OverdueSearchModel, OverdueSearchEffect> {
+class OverdueSearchInit : Init<OverdueSearchModel, OverdueSearchEffect> {
 
   override fun init(model: OverdueSearchModel): First<OverdueSearchModel, OverdueSearchEffect> {
-    val effects = mutableSetOf<OverdueSearchEffect>()
-    if (isOverdueSearchV2Enabled) {
-      effects.add(LoadVillageAndPatientNames)
-    } else {
-      effects.add(LoadOverdueSearchHistory)
-    }
-
-    effects.add(LoadSelectedOverdueAppointmentIds)
-
-    return first(model, effects)
+    return first(model, setOf(LoadVillageAndPatientNames, LoadSelectedOverdueAppointmentIds))
   }
 }