+++ app/src/main/java/org/simple/clinic/home/overdue/search/OverdueSearchModel.kt
@@ -1,27 +1,18 @@
 package org.simple.clinic.home.overdue.search
 
 import android.os.Parcelable
-import androidx.paging.PagingData
 import kotlinx.parcelize.IgnoredOnParcel
 import kotlinx.parcelize.Parcelize
-import org.simple.clinic.home.overdue.OverdueAppointment
 import java.util.UUID
 
 @Parcelize
 data class OverdueSearchModel(
-    val overdueSearchHistory: Set<String>?,
-    val searchQuery: String?,
     val overdueSearchProgressState: OverdueSearchProgressState?,
-    @IgnoredOnParcel
-    val overdueSearchResults: PagingData<OverdueAppointment> = PagingData.empty(),
     @IgnoredOnParcel
     val selectedOverdueAppointments: Set<UUID> = emptySet(),
     val villageAndPatientNames: List<String>?,
     val searchInputs: List<String>
 ) : Parcelable {
-
-  val hasSearchQuery: Boolean
-    get() = !searchQuery.isNullOrBlank()
 
   val hasSearchInputs: Boolean
     get() = searchInputs.isNotEmpty()
@@ -33,8 +24,6 @@ data class OverdueSearchModel(
 
     fun create(): OverdueSearchModel {
       return OverdueSearchModel(
-          overdueSearchHistory = null,
-          searchQuery = null,
           overdueSearchProgressState = null,
           selectedOverdueAppointments = emptySet(),
           villageAndPatientNames = null,
@@ -43,22 +32,8 @@ data class OverdueSearchModel(
     }
   }
 
-  fun overdueSearchHistoryLoaded(
-      overdueSearchHistory: Set<String>
-  ): OverdueSearchModel {
-    return copy(overdueSearchHistory = overdueSearchHistory)
-  }
-
-  fun overdueSearchQueryChanged(searchQuery: String): OverdueSearchModel {
-    return copy(searchQuery = searchQuery)
-  }
-
   fun loadStateChanged(overdueSearchProgressState: OverdueSearchProgressState): OverdueSearchModel {
     return copy(overdueSearchProgressState = overdueSearchProgressState)
-  }
-
-  fun overdueSearchResultsLoaded(overdueSearchResults: PagingData<OverdueAppointment>): OverdueSearchModel {
-    return copy(overdueSearchResults = overdueSearchResults)
   }
 
   fun selectedOverdueAppointmentsChanged(selectedOverdueAppointments: Set<UUID>): OverdueSearchModel {