+++ app/src/main/java/org/simple/clinic/home/overdue/search/OverdueSearchUi.kt
@@ -1,9 +1,7 @@
 package org.simple.clinic.home.overdue.search
 
 interface OverdueSearchUi {
-  fun showSearchHistory(searchHistory: Set<String>)
   fun hideSearchResults()
-  fun hideSearchHistory()
   fun hideProgress()
   fun showProgress()
   fun showSearchResults()