+++ app/src/main/java/org/simple/clinic/home/overdue/search/OverdueSearchUiActions.kt
@@ -7,14 +7,12 @@ import java.util.UUID
 interface OverdueSearchUiActions {
   fun openPatientSummaryScreen(patientUuid: UUID)
   fun openContactPatientSheet(patientUuid: UUID)
-  fun setOverdueSearchQuery(searchQuery: String)
   fun openSelectDownloadFormatDialog()
   fun openSelectShareFormatDialog()
   fun openShareInProgressDialog()
   fun showNoInternetConnectionDialog()
   fun setOverdueSearchResultsPagingData(
       overdueSearchResults: PagingData<OverdueAppointment>,
-      selectedOverdueAppointments: Set<UUID>,
-      searchQuery: String
+      selectedOverdueAppointments: Set<UUID>
   )
 }