+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Add overdue search query for supporting multiple search inputs
 - Fix overdue search history and results not showing in overdue search screen
 - Add multiple search inputs support in overdue search screen
+- Remove overdue search v1 implementation
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 
 ## 2022-08-01-8350